Package esdb_topo.py as a part of oscars-backend.
There are a couple sub-optimal aspects of this change:

1.  esdb_topo.py depends on the requests module, which is an additional
package that needs to be installed, but there is no dependency
information created to force installation of requests.

2.  esdb_topo.py depends on some of the scripts/libraries in
neg-tools/esnet-python for its functionality.  We don't have an
easy way of expressing a dependency on this, since we might not
have an RPM packaging for esnet-python.

3.  There are probably some default configuration bits that need to be
frobbed to really consider this part done.

Towards #135.