fixed Syslogger bean not found issue (by duplicating the class to both backend and PSS)
why it didn't work? the pss module is configured to look for beans in the net.es.oscars.pss package and below (as per spring boot defaults and PssApp.java); Syslogger.java was not under that package and therefore could not be found

another approach would be to "by hand" add it to PssApp's configuration; I prefer the class duplication to messing around with that especially for what is basically a placeholder class.